FIXUP: spi: spi-geni-qcom: Add SPI driver support for GENI based QUP (prepare_transfer_hardware)
The prepare_transfer_hardware() function is supposed to run code to
prepare each transfer.  It was being used instead to do one-time init
that belonged in probe.  Move it to probe.  This has to happen with
pm_runtime working so we can access hardware registers, but that's
fine.

It is intended that this patch be squashed into the patch it fixes to
become a new revision of that patch.

BUG=b:80592084
TEST=Use cros_ec a whole bunch

Change-Id: Icecddac42abe214b6b696d322e994007f1f60164
Signed-off-by: Douglas Anderson <dianders@chromium.org>