FIXUP: spi: spi-geni-qcom: Add SPI driver support for GENI based QUP (__relaxed)
Gratuitously using the __relaxed versions of readl() and writel() is
discouraged in Linux because it gives very little performance gain in
99.9% of the cases and it makes the code significantly harder to
reason about it.  Remove it in spi-geni-qcom.  If it is shown that a
particular read or write is in a performance critical section then
that one instance can be added back in after carefully analysis and
documentation.

It is intended that this patch be squashed into the patch it fixes to
become a new revision of that patch.

BUG=b:80592084
TEST=Use cros_ec a whole bunch

Change-Id: I84a6d21729ad371116a619166e170a128e7510ec
Signed-off-by: Douglas Anderson <dianders@chromium.org>