CHROMIUM: config: Normalize (PPTP VPN / PPP / NAT stuff)
Run "./chromeos/scripts/kernelconfig olddefconfig" to normalize
things.  This is part of the result.

This is pretty much all fallout of the non-normalized CL:1157955.
Specifically:

For "CONFIG_NF_CT_PROTO_GRE=y" and "CONFIG_NF_NAT_PROTO_GRE=y"
=> CL:1157955 has "CONFIG_NF_NAT_PPTP=y".  That selects
  "NF_NAT_PROTO_GRE".
=> CL:1157955 also has "CONFIG_NF_CONNTRACK_PPTP=y".  That selects
  "NF_CONNTRACK_PPTP".

--

For removals of "CONFIG_PPPOLAC=y" and "CONFIG_PPPOPNS=y":
=> CL:1157955 introduced them.  These configs don't actually exist in
   Chrome OS, so they just fall out.

--

For "CONFIG_SLHC" changing from "=m" to "=y":
=> CL:1157955 changed CONFIG_PPP from =m to =y and SLHC is selected by
   PPP.

===

Presumably someone should double-check all this stuff since it implies
that the original cherry-pick might not have been doing exactly what
people thought it was.  ...but in the meantime we should land this to
get the config normalized.

BUG=chromium:207904, b:112024433
TEST=None

Change-Id: I20123aa9c3a163216890a587ea40457e2b093c2a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1204233
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>