HACK: WIP: drm/bridge: ti-sn65dsi86: magic delay 70 => 100
We don't know why but we need this magic delay to be bigger.  With 70
ms on my board I saw about a 25% failure in link training.  Increasing
this delay (or putting in the hack to keep pp3300_dx_edp hack always
on) fixed it.  It seems better to just enable the delay for now.

TODO: root cause this.  The rail we were leaving on is one that goes
straight to the panel, so maybe we are somehow violating timing
parameters on the panel???

BUG=b:112356283
TEST=Reboot test, no link training failures

Change-Id: Ifb2c08541a3c1c285c8e03e94da9a1cfd4c7834f
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1213617