CHROMIUM: "nosymfollow" mount option for chromiumos LSM
Add a custom option that can be passed to the mount() call which will
cause the chromiumos LSM to enforce that symlinks are not traversed on
the given file system. This will allow CrOS to disallow symlink
traversal on user-supplied file systems.

BUG=chromium:867807
TEST=autotest in CL:1269737

Change-Id: I9ed266e82102b6b0422da94c037eb93c16fa1c58
Signed-off-by: Micah Morton <mortonm@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1217762
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>