CHROMIUM: config: Normalize (CONFIG_DRM_AMDGPU_IP_BLOCK_MASK)
Run "./chromeos/scripts/kernelconfig olddefconfig" to normalize
things.  This is part of the result.

+CONFIG_DRM_AMDGPU_IP_BLOCK_MASK=0xffffffff
 From: commit c2173b35c36b ("FROMLIST: drm/amdgpu: add ip_block_mask
user option for static builds"). CL:1184599 add a value for the config,
but is later reverted by CL:1215511, which cause the config not having
a default value.

BUG=None
TEST=None

Change-Id: Ie3e26c1adf8e7c76b3980041db79bc15498884fb
Signed-off-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1220976
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>