CHROMIUM: arm64: move BP hardening to check_and_switch_context
This is an backport of upstream a8e4c0a919ae

ASID and mm context handling was re-worked significantly after 3.18 and the
check_and_switch_context function was considerably different and lived
in arch/arm64/include/asm/mmu_context.h. The mm switch itself could
happen in a few different places, so we need to apply bp hardening in all
of those spots.

BUG=chromium:889539
TEST=build/boot on hana with USE=kvm_host
 - ensure SMC workaround is being called

Change-Id: I5e856ccbee8c50921254938971637502f8916afb
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1245254