CHROMIUM: Revert "drm/i915: Warn against variable length arrays"
This reverts commit c5c2b11894f4f862cf243b955ac59bb1a5fe61b9.

The changes enable a warning about variable length arrays for the i915
driver. While this may be generally desirable it results in a flood
of warnings about VLAs in __compiletime_error_fallback when building
with clang.

If the error condition is true__compiletime_error_fallback() generates
an array with a negative size to force a build error. __compiletime_assert()
assigns the error condition to a variable and passes it to
__compiletime_error_fallback(), which uses it to compute the size of the
array. With the variable clang can't determine the constness of the
condition, and hence emits a VLA and the corresponding warning.

mka@: The FROMLIST patch "compiler.h: give up __compiletime_assert_fallback()"
(https://lore.kernel.org/patchwork/patch/977668/) is a possible
alternative to the revert. If it lands we should definitely pick it
instead of reverting, otherwise the revert is probably the option with
less potential conflicts from merges.

BUG=chromium:891028
TEST=build for caroline

Change-Id: Ia7d979b89550fc930181109b68180c87532b0983
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>