CHROMIUM: ASoc: dmic: add modeswitch delay
Dual Mode Mic's need some delay to complete the mode switch. The mode
switch starts as soon as the CLK starts running. To complete the mode
switch the CLK should be running for the mode switch delay which is
defined by the DMIC data sheet. The mode switch delay parameter may be
defined in the ACPI tables.

BUG=b:112888584
TEST=verify DMIC record

Change-Id: Ic8c8be1aa04e8d35ad4827d81a37ee54c25dfc63
Signed-off-by: Sathyanarayana Nujella <sathyanarayana.nujella@intel.com>
Signed-off-by: Jairaj Arava <jairaj.arava@intel.com>
Signed-off-by: Harsha Priya <harshapriya.n@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1250541
Reviewed-by: Benson Leung <bleung@google.com>