CHROMIUM: HID: google: whiskers: more robust tablet mode detection
The USB interface may get detected before the platform/EC one, so let's
note the state of the base (if we receive event) and use it to correctly
initialize the tablet mode switch state.

Also let's start the HID interface immediately when probing, this will
ensure that we correctly process "base folded" events that may be sent
as we initialize the base.

BUG=b:117439202
TEST=Build and boot Nocturne, make sure tablet mode is recognized on
startup with Whiskers in folded state.

Change-Id: I873d0e64739d1fb1390d50bf99986cd147b6e708
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1271261
Reviewed-by: Benson Leung <bleung@chromium.org>