CHROMIUM: media: rockchip-vpu: Honor format number of planes in encoder
Currently the encoder always assumes 3 planes, but we also advertise
support for 1- and 2-plane formats, so honor them.

This fixes a null pointer dereference which happened when the driver
tries to get a DMA address of an out of bounds plane.

While at it, remove dummy encode handling from RK3399 encoder code,
since the workaround is not needed on this SoC and it's already
disabled in higher level code.

BUG=chromium:893530
TEST=Run VEA unit test with NV12 input on Kevin

Change-Id: Ic34813d84a68b1e786b009ebf26de1f271e3b61c
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1276105
Commit-Ready: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>