CHROMIUM: iwl7000: mvm: make NVM access actually fail on failures
On any failure, including if we crash the firmware or get garbage
data, we currently ignore this and pretend the OTP was empty.
Clearly, this isn't typically the case.

In cases other than the firmware saying it can't read the requested
section, or the section having ended, make the access actually fail
and trickle the error up through the layers to fail init.

hw=all

Change-Id: I754828f975400a84eca78570b1cc17b791a92e6c
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: b63f221ae492b07e627db07448910eb9d28a6739