CHROMIUM: iwl7000: chromeOS: make pertid an array in < 4.18
In kernels older than 4.18, we assume that pertid in the sinfo struct
is a preallocated array, but it is not, because the entire sinfo
struct is backported.  This causes a NULL pointer dereference when
trying to use it.  Fix it by making it an actual array in < 4.18
cfg80211's.

Change-Id: I9d5b07b43e669e8f8ca259bc534c4dce6d165bec
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: b54c1571f5b7e2bfe5f14b6d56fdd41d4645f2da