CHROMIUM: drm/i915: Fix divide-by-zero Oops
i915 readout doesn't fill all of the various state out and assumes that
the first atomic commit it receives is going to be a modeset. In some
cases (for R67, it's all cases) Chrome will send a non-modeset commit to
the kernel.

When the non-modeset commit is received, intel_state is not filled in
and has intel_state->active_crtcs == 0. Since readout has determined
one crtc is active, we try to calculate the watermark which involves
dividing by the number of active crtcs. Since intel_state->active_crtcs
== 0, we Oops.

This patch fixes exactly that scenario and assumes the rest of the
driver will cope with the inconsistent state until we get a modeset.
I've got a more general fix for this upstream, but it is much more
invasive and given the drift between 3.18 and 4.19, is not guaranteed to
work as well here as it does upstream.

BUG=chromium:878595
TEST=Tested on asuka. Use crosreview.com/1284610 for automatic testing
     on R70. Ran through 4150 successful cycles with lid closed, no
     crashes.

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I56a7572674497b8a6c08420a2a68d6d3e26ce91c
Reviewed-on: https://chromium-review.googlesource.com/1281782
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>