CHROMIUM: cros_ec: Remove wake capability from cros-ec-ctl devices.
ACPI attaches the firmware node of the parent to the platform
devices. If the parent firmware node has a valid wakeup flag, ACPI
marks this platform device also as wake capable.  But this platform
device by itself cannot wake the system up. Mark the wake capability to
false. This is no-op if the platform device is not wake capable in the
first palce.

BUG=b:112974410
TEST=manual, run on nocturne to test powerd recognizes that
fingerprint is the wake source.

Change-Id: I7224e230c260be6f14e97c4486f0f4d332728b23
Signed-off-by: RaviChandra Sadineni <ravisadineni@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1297833
Commit-Ready: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>