CHROMIUM: USB: quirks: add NO_LPM quirk for Logitech Rally Camera
Description: Some USB device / host controller combinations seem to have
problems with Link Power management. In particular it is described that
the combination of a Logitech Rally and Atrus device causes
'not enough bandwidth for new device state'error

This patch creates quirk entries for the Logitech Rally Camera
indicating LPM should remain disabled for the device.

Note: This is an established workaround for an ongoing partner bug with
the specific intent of getting the Rally Camera working on Teemo
@see b/65034294

TEST=While using Logitech Rally camera w/ Atrus on Teemo and Buddy

BUG=chromium:901469, b:115756398

Change-Id: I67d9943cd5d66604b8bbee89f2ce240db8f7cb04
Signed-off-by: Kyle Williams <kdgwill@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1297535
Commit-Ready: Kyle Williams <kdgwill@chromium.org>
Tested-by: Kyle Williams <kdgwill@chromium.org>
Tested-by: Andrew Wong <wongan@google.com>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Andrew Wong <wongan@google.com>