CHROMIUM: lib/lzo: implement run-length encoding
When using zram, we frequently encounter long runs of zero bytes.
This adds a special case which identifies runs of zeros and encodes
them using run-length encoding.

This is faster for both compression and decompresion. For
high-entropy data which doesn't hit this case, impact is minimal.

Compression ratio is within a few percent in all cases.

This modifies the bitstream in a way which is backwards compatible
(i.e., we can decompress old bitstreams, but old versions of lzo
cannot decompress new bitstreams).

BUG=chromium:907138
TEST=python -c 'l = [0] * 1024*1024*400; l2 = [0] * 1024*1024*400'
TEST=zramctl; dmesg
TEST=zramctl should show evidence of swapping using the lzo algorithm,
TEST=and dmesg should not report any errors

Change-Id: I6f33a481483dd7dc9462e076f92b1827dd252c69
Signed-off-by: Dave Rodgman <dave.rodgman@arm.com>
Reviewed-on: https://chromium-review.googlesource.com/1307378
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>