CHROMIUM: media: intel-ipu3: Adjust order when queuing buffers to css
For imgu, FW starts processing when input buffer is queued.
When starts, FW requires all output buffers to be queued
and ready to use. The current logic queues input buffer first
and then followed by the other output buffers. This gives the
chance that when FW starts, not all output buffers are queued
to it which causes imgu hang.
This patch adjust the order of buffer queuing and make sure
input buffer is queued last.

BUG=b:118591974
TEST=verify basic camera functions with Camera Nexus app

Change-Id: I3a6425360bba8a3501e964a660ce090e13e4a120
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1303374
Reviewed-by: Tomasz Figa <tfiga@chromium.org>