CHROMIUM: iwl7000: mvm: add an option to dereference vif by id
Currently whenever we get firmware notification with mac id,
we iterate over all the interfaces to find the ID. This is a
bit cumbersome. Instead, adding an array of RCU pointers, like
we have for station IDs. This is not expensive space wise
since we have only up to 4 active MACs, and not complicated
code wise, since we have a clear point to init and de-init it.

Change-Id: I8c688e2cabb0698289161d63de903332f3ae39fc
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: a7ee2bc304b2a89bfb34e17e0e3763d6af45efaf