HACK: CHROMIUM: media: intel-ipu3: Remove unnecessary checking for parameter pool
imgu is working frame by frame and the driver only uses
the latest 2 parameter sets. We now have a pool depth of 4
so overwriting the oldest one should always be safe.
The checking for pool status is unnecessary and is making
false result at the beginning and every 4 frames which
causes parameters not correctly set. So remove this checking.

BUG=b:116179323
TEST=verify basic camera functions with Camera Nexus app

Change-Id: Ifaac26a98412029a89d572dd98f5bb3b0226a7a4
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1309674
Reviewed-by: Tomasz Figa <tfiga@chromium.org>