HACK: CHROMIUM: rk3399-vpu: Compute plane address with data_offset
With V4L2 MEMORY DMABUF API, it is necessary to set a file descriptor on
each plane. The way to pass an offset with a DmaBuf is not defined in V4L2
specification. This is the case as it is needed to specify offsets because
Chromium creates one DmaBuf for the planes.
This is HACK to abuse data_offset for the purpose.
This patch should be reverted after the appropriate way is defined.

BUG=chromium:895230, chromium:901264
TEST=VEA unittest --native_input on kevin

Change-Id: I0508ac10f2c2be4a0332125f41d0dd4b13d0499c
Signed-off-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1312535
Reviewed-by: Tomasz Figa <tfiga@chromium.org>