ASoC: Intel: correct codec PLL configuration and power management by codec
The sysclk should come from PLL which is proper frequency, not MCLK.
Moreover, the source of sysclk should be decided before DAPM of codec.
The source of sysclk doens't need to set back to MCLK. Because codec driver will
use this condition to decide some power turning on or off.

BUG=b:118841107
TEST= playing music via Headphone plugged-in and listening the sound is clearly without noise

Change-Id: I10098f433382d7ca096c6a61d9d96b5259af6309
Signed-off-by: Shuming Fan <shumingf@realtek.com>
Signed-off-by: Mac Chiang <mac.chiang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/1358280
Reviewed-by: Marco Chen <marcochen@chromium.org>
Commit-Queue: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>