HACK: CHROMIUM: mtk-vcodec: Compute plane address with data_offset
With V4L2 MEMORY DMABUF API, it is necessary to set a file descriptor on
each plane. The way to pass an offset with a DmaBuf is not defined in V4L2
specification. This is the case as it is needed to specify offsets because
Chromium creates one DmaBuf for the planes.
This is HACK to abuse data_offset for the purpose.
This patch should be reverted after the appropriate way is defined.

BUG=chromium:895230, chromium:901264
TEST=VEA unittest --native_input on hana

Change-Id: I7ad238d9bdb0916d429720faef27783fffeb2069
Signed-off-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1319829