HACK: CHROMIUM: rk3288-vpu: Compute plane address with data_offset
With V4L2 MEMORY DMABUF API, it is necessary to set a file descriptor on
each plane. The way to pass an offset with a DmaBuf is not defined in V4L2
specification. This is the case as it is needed to specify offsets because
Chromium creates one DmaBuf for the planes.
This is HACK to abuse data_offset for the purpose.
This patch should be reverted after the appropriate way is defined.

BUG=chromium:895230, chromium:901264
TEST=VEA unittest --native_input on veyron_minnie

Change-Id: Id1471ef622aa3e616dcdb5559f4f948b2c1155ee
Signed-off-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1319830
Reviewed-by: Tomasz Figa <tfiga@chromium.org>