CHROMIUM: iio: cros_ec_sensors: track wakeups based on _WAKEUP flag
There are a few cases where cros_ec sensors might want to wake the
AP -- particularly with a "significant motion" event. Currently, the
main cros_ec device treats all EC_MKBP_EVENT_SENSOR_FIFO interrupts as
wakeup events in its ISR, but this is not the right way to do it. There
are plenty of non-wakeup sensor events that occur, and if we treat them
as wakeups, this may disturb the wakeup_count and interrupt entry into
system suspend (e.g., S3).

So instead, let's log the wakeup event in the cros_ec_sensors driver,
where we parse the MOTIONSENSE flags. After we do this, it's safe to
start ignoring EC_MKBP_EVENT_SENSOR_FIFO events in the parent device,
with respect to wakeups.

BUG=b:118685087
TEST=suspend_stress_test on scarlet; find a device that supports
     "significant motion" and ensure it still increases 'wakeup_count'

Change-Id: Ibc162fe4d0fc335d525595e27d2d377cd4104e51
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1319994
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>