CHROMIUM: cros-ec: don't report EC_MKBP_EVENT_SENSOR_FIFO as wakeup
[Adapted for upstream here:
https://lore.kernel.org/patchwork/patch/1008110/
https://lkml.kernel.org/r/20181108024939.174321-2-briannorris@chromium.org
but much of this was refactored and moved to drivers/platform/chrome/,
so I don't bother with a direct BACKPORT.
Note that patch 1 of 2 in that series doesn't really apply to this
kernel, since we don't have the incorrect documentation here, and we
were already ignoring 'ret == 0' cases from cros_ec_get_next_event().]

EC_MKBP_EVENT_SENSOR_FIFO events can be triggered for a variety of
reasons, and there are very few cases in which they should be treated as
wakeup interrupts (particularly, when a certain
MOTIONSENSE_MODULE_FLAG_* is set, but this is not even supported in the
mainline cros_ec_sensor driver yet). Most of the time, they are benign
sensor readings. In any case, the top-level cros_ec device doesn't know
enough to determine that they should wake the system, and so it should
not report the event. This would be the job of the cros_ec_sensors
driver to parse.

This patch adds checks to cros_ec_get_next_event() such that it doesn't
signal 'wakeup' for events of type EC_MKBP_EVENT_SENSOR_FIFO.

This patch is particularly relevant on devices like Scarlet (Rockchip
RK3399 tablet, known as Acer Chromebook Tab 10), where the EC firmware
reports sensor events much more frequently. This was causing
/sys/power/wakeup_count to increase very frequently, often needlessly
interrupting our ability to suspend the system.

BUG=b:118685087
TEST=suspend_stress_test on scarlet; find a device that supports
     "significant motion" and ensure it still increases 'wakeup_count'

Change-Id: Ia008078f59e486de88bb8d429b2eb171f6be056b
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1319996
Reviewed-by: Enrico Granata <egranata@chromium.org>
(cherry picked from commit 6322a39140dcb2f7e3bcc5c03056ce3abfd33b5c)
Reviewed-on: https://chromium-review.googlesource.com/c/1343770