CHROMIUM: iio: cros_ec: enable spreading across sensor batches
Sensor timestamps can end up being "in the past" and, after CL:1335860,
there is no fail-safe clamping them. While intuitively appealing, it is
not feasible nor correct to always mark timestamps in the past as being
equal to the "newest even seen", even if done on a per-sensor basis.

The sensor ring has a way to handle multiple sensor samples with an
identical timestamp, and that is the spreading logic. However, that
logic only works in a "batch", and if multiple samples with identical
timestamp show up in different batches, it wouldn't spread them.

This patch is an attempt at improving that situation by introducing
spreading across sensor batches.

It works by enabling the spreading logic to "reopen" a batch that
was already pushed out, and adding new samples to it after the fact.

BUG=b:117169255
TEST=observe timestamps spread across batches

Change-Id: I3896b9dba8c6bee0671a447cd3809672fda6ee57
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1336931
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>