CHROMIUM: iio: cros_ec: enable spreading across sensor batches
Sensor timestamps can end up being "in the past" and, after CL:1335860,
there is no fail-safe clamping them. While intuitively appealing, it is
not feasible nor correct to always mark timestamps in the past as being
equal to the "newest even seen", even if done on a per-sensor basis.

The sensor ring has a way to handle multiple sensor samples with an
identical timestamp, and that is the spreading logic. However, that
logic only works in a "batch", and if multiple samples with identical
timestamp show up in different batches, it wouldn't spread them.

This patch is an attempt at improving that situation by introducing
spreading across sensor batches.

It works by enabling the spreading logic to "reopen" a batch that
was already pushed out, and adding new samples to it after the fact.

This CL is an aggregate of several CLs landed on chromeos-4.4 kernel:
  CL:1376689
  CL:1370689
  CL:1376769
  CL:1389076

BUG=b:117169255
TEST=observe timestamps spread across batches

Change-Id: Ic36cbe9f59fa1328c89ce0560312beca833b9b03
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1376689
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
(cherry picked from commit 179a370498d13b7ee193d8a314a98a5d156cb56a)
Reviewed-on: https://chromium-review.googlesource.com/c/1412833
Reviewed-by: Gwendal Grignou <gwendal@google.com>
Commit-Queue: Gwendal Grignou <gwendal@google.com>
Tested-by: Gwendal Grignou <gwendal@google.com>