CHROMIUM: cros_ec: handle MKBP more events flag
CL:1341159 adds a "more events" flag to allow the EC to notify us that
there are more MKBP events waiting to be served. This CL extends the
cros_ec bottom-half IRQ handler to listen to that flag in order to try
and serve all known events before returning to the kernel with IRQ_HANDLED.

This serves an optimization purpose as it minimizes the round-trip back
and forth from the interrupt handler, and it also improves the quality
of sensor timestamping, as it avoids a mismatch of event times between
EC and AP that can confuse the median filter we use in the sensor ring.

As described in the EC CL, this change is forward and backward compatible.

BUG=b:119570064, b:112102925
TEST=on nocturne, enable sensors and use hardware buttons and
     observe the kernel receive MKBP events with the flag enabled

Change-Id: I1a924bfe0255367329f7a07d4e2b62af2709efa1
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1340983
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>