CHROMIUM: media: mtk-vcodec: Remove VA from encoder frame buffers
The encoder driver has no need to do any CPU access to the source frame
buffers. Use a separate structure for holding DMA addresses and sizes
without a field for VA for those, so we do not end up introducing any
erroneous dereferences of the VAs in the old struct.

BUG=chromium:906473
TEST=emerge-hana chromeos-kernel-3_18 (no dereferences of the removed
     field)

Change-Id: I48e59b43a587f87fbb1cdbea2c7357a78a320b80
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1341646
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>