BACKPORT: FROMGIT: arm64: dts: rockchip: Add 32k clk on rk3399-gru
This adds the 32k clock to the RK3399 Gru board file, which is provided
by a Silego oscillator on Gru boards.

Even though it's not directly used, muxes will end up traversing the
entire clk tree on calls to determine_rate if it doesn't exist. This
is because the 32k clk is listed as a possible parent on some clks.
Since the clk doesn't know about the 32k clk (it was never registered),
it triggers a global search for it. This can happen about 40 times per
second, which isn't great for power.

Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
[moved clock position and adapted commit message a bit]
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
(cherry picked from commit 17222eb932ad2ec67e583179f507982460d81bde
 https://git.kernel.org/pub/scm/linux/kernel/git/mmind/linux-rockchip.git/
 for-next)

Conflicts:
	[Context delta]
	arch/arm64/boot/dts/rockchip/rk3399-gru.dtsi

BUG=b:120160928
TEST=boot kevin

Change-Id: Id52c4bf5ef569e2158b7c7b2ddc07d6018ec7f92
Reviewed-on: https://chromium-review.googlesource.com/1354522
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>