CHROMIUM: iwl7000: mvm: Fix handling of Rx no data indication
When an Rx no data indication is received, an SKB is allocated
to hold the radiotap information without holding an actual frame
data. When the frame is to be sent to mac80211, the code path
tries accessing the SKB assuming that the buffer holds a frame,
which is wrong.

Fix this by skipping data path handling in such cases.

Change-Id: I118b4614045eb70afd5cf14498c44adc2b84aa7e
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: f9176d30cb28798d607ce0c168531984b5d939b1