Revert "drm/rockchip: Allow driver to be shutdown on reboot/kexec"
This reverts commit 67dbeda8f4bb1c728683869ee7c35d50f73e76a3.

The rockchip display driver is not prepared for this, at least on our
variant of "Linux 4.4". There are a variety of crashes in the MIPI and
eDP variants of this driver when performing shutdown. Note that these
crashes also may happen if we unbind/remove the driver, like:

  echo display-subsystem > /sys/bus/platform/drivers/rockchip-drm/unbind

But that doesn't happen in practice.

BUG=b:120498192
TEST=power off kevin, scarlet -- watch kernel log, ensure no crashes or
     warnings

Change-Id: I04204a8c39304a4e94c08cd9661cf8f778a82035
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1361611
Reviewed-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit d0622d0f77f0fa6817770185b1c51dcd61116648)
Reviewed-on: https://chromium-review.googlesource.com/c/1367940