UPSTREAM: ALSA: hda/hdmi: Always set display_power_control for Intel HSW+ codecs
We've excluded the display_power_control flag for Intel HSW and BDW
codecs as the HD-audio controllers of the corresponding platforms take
care of the display power as well.  But the recent refactoring
separates the controller and the codec power accounting, so it's fine
to call the display PM even for HSW/BDW codecs.  This is less
confusing since we can avoid this well-hidden condition.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 46594d3345f0432ddc83b9f3c1b04492fdfafd07)

BUG=b:130538528
TEST=Test Audio use cases for GLK with full SOF patch series applied.

Change-Id: I80855b6218574fee1f33d21148d9711d71127cb8
Signed-off-by: smit shah <smit.shah@intel.corp-partner.google.com>
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>