CHROMIUM: iio: cros_ec: Do not prevent cross-batch spreading
The code used to mark sensor timestamps for spreading was using a condition
that cast too wide a net and would result in samples with the same "a"
value being marked in such a way that the spreading code would fail to
recognize them as part of the same batch and cause samples to be lost.

One such example collected during debugging goes as follows:
the last sensor batch timestamp was 121683256111
and then a set of events came in for which a=273139352

The filter placed the first one at 123791054426, and upon trying to
place the following ones at the same value, the code would detect
that 123791054426 >= 123791054426 and move the sample back to the
previous batch's timestamp 121683256111. This rollback of time would
in turn cause the spreading code to get confused and drop samples.

The fix here is to leave samples with the same timestamp alone and let
the spreading handle them organically, instead of trying to force them
back in time.

BUG=b:120508077
TEST=observe sensor timestamps being spread on bobba

Change-Id: I96731d7ec420e4182e5f2f817b836cbd4f8e02ad
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1370689
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>