CHROMIUM: iwl7000: mac80211: support max channel switch time element
2018 REVmd of the spec introduces the max channel switch time
element which is optionally included in beacons/probes when there
is a channel switch / extended channel switch element.
The value represents the maximum delay between the time the AP
transmitted the last beacon in current channel and the expected
time of the first beacon in the new channel, in TU.

Parse the value and pass it to the driver.

Change-Id: I3dd3a829391f9b2f8453c088f6455ec3ae2c5b99
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: d0d22b675c612bf86fa285e314c7c15d7a59d3dd