INTERNAL: Bluetooth: btqca: inject command complete event during fw download.
Latest qualcomm chips are not sending an command complete event for
every firmware packet sent to chip. They only respond with a vendor
specific event for the last firmware packet. This optimization will
decrease the BT ON time. Due to this we are seeing a timeout error
message logs on the console during firmware download. Now we are
injecting a command complete event once we receive an vendor specific
event for the last RAM firmware packet.

Signed-off-by: Harish Bandi <c-hbandi@codeaurora.org>

BUG=b:121286158
Signed-off-by: Sumit Agrawal <agsumit@codeaurora.org>

Change-Id: I999c79871bedb21bd53a91ca6544c0e13f8d1aa3
Reviewed-on: https://chromium-review.googlesource.com/c/1394468
Reviewed-by: Suresh Rajashekara <sureshraj@chromium.org>
Reviewed-by: Zhihong Yu <zhihongyu@chromium.org>
Commit-Queue: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>