CHROMIUM: iwl7000: pcie: fix emergency path
Allocator swaps the pending requests with 0 when it starts
working. This means that relying on it n RX path to decide if
to move to emergency is not always a good idea, since it may
be zero, but there are still a lot of unallocated RBs in the
system. Change allocator to decrement the pending requests on
real time. It is more expensive since it accesses the atomic
variable more times, but it gives the RX path a better idea
of the system's status.

Change-Id: I7e2a509d9a2765d332e4d18e7c9ea1e438fe47ed
Reported-by: Ilan Peer <ilan.peer@intel.com>
Signed-off-by: Sara Sharon <sara.sharon@intel.com>

[BUGFIX] iwlwifi: pcie: Fix tracking of pending RB requests

Commit 2dadcc5dd8d9 ("[BUGFIX] iwlwifi: pcie: fix emergency path")
changed the allocator to track the pending requests in real time, but
introduced an inconsistency in the tracking in case the local 'pending'
variable reached 0. In this case, 'pending' is set again to the value
of 'rba->req_pending', and thus, pending is 'incremented' again, resulting
with another iteration in the loop, which in tern causes another atomic
decrement of 'rba->req_pending' (which can be become negative). From this
point on the pending request tracking is broken.

Fix this by moving the atomic_dec operation before decrementing the local
'pending' variable, and before handling the case of 'pending == 0'.

Change-Id: I98c36e7ba6bda9ff7f44a942e0b5a7dd4fde4a7a
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: d93a1c981818b472f403ad451203d7788f45ffcd