CHROMIUM: iwl7000: pcie: Fix tracking of pending RB requests
Commit 2dadcc5dd8d9 ("[BUGFIX] iwlwifi: pcie: fix emergency path")
changed the allocator to track the pending requests in real time, but
introduced an inconsistency in the tracking in case the local 'pending'
variable reached 0. In this case, 'pending' is set again to the value
of 'rba->req_pending', and thus, pending is 'incremented' again, resulting
with another iteration in the loop, which in tern causes another atomic
decrement of 'rba->req_pending' (which can be become negative). From this
point on the pending request tracking is broken.

Fix this by moving the atomic_dec operation before decrementing the local
'pending' variable, and before handling the case of 'pending == 0'.

Change-Id: I98c36e7ba6bda9ff7f44a942e0b5a7dd4fde4a7a
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: 10aec54868cc71f4235c6eee41bf64419b58b469