CHROMIUM: iwl7000: mvm: fix queue stop flow
The queue stop flow references to queue_info, which is not updated
at all for 22000 devices. This results with bad station id, and a
bail out of the function. Unfortunately, this is a silent error flow,
and wasn't noticed so far.

There are almost no side effects, since packets just go to overflow
queue in the transport. However, it is an abuse of the overflow queue
intention and a nullification of the start and stop flows. In addition,
it exposed a race between the flush flow and the overflow queue (which
should be fixed separately).

Add station id to TVQM info and use it.

Change-Id: I048b7fe34d7f98ef8c5499468b79f058fe01d70b
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: 9844a4ef1043323144d78b555a6fa2e16cd5b763